+++ morpheus-graphql-core/src/Data/Morpheus/Core.hs
@@ -24,6 +24,7 @@ module Data.Morpheus.Core
     RenderGQL (..),
     render,
     ValidateSchema (..),
+    internalSchema,
   )
 where