+++ morpheus-graphql-core/changelog.md
@@ -4,6 +4,8 @@
 
 ### New features
 
+- `Data.Morpheus.Core` provides default GrapHQL type definitions with `internalSchema`
+- exposed `Data.Morpheus.Internal.Ext`
 - exposed `Data.Morpheus.Types.GQLWrapper` with `EncodeWrapper` and `DecodeWrapper` type-classes.
 
 ### Breaking changes
@@ -32,8 +34,6 @@
     to `ByteString -> Eventless [TypeDefinition ANY CONST]`
 
 ### new features
-
-- `Data.Morpheus.Core` provides default GrapHQL type definitions with `internalSchema`
 
 ### Minor Changes