+++ morpheus-graphql-core/changelog.md
@@ -10,6 +10,8 @@
 
 ### Breaking changes
 
+- `parseTypeSystemDefinition` and `parseGQLDocument` is replaced with `parseSchemaDocument`
+
 - removed `parseDSL` from `Data.Morpheus.Core`
 
 - following Types and modules are migrated to the new package `morpheus-graphql-app`: