+++ morpheus-graphql-core/changelog.md
@@ -10,8 +10,8 @@
 
 ### Breaking changes
 
-- `parseTypeSystemDefinition` and `parseGQLDocument` is replaced with `parseSchemaDocument`
-- `parseFullGQLDocument` replaced with `parseFullSchemaDocument`
+- `parseTypeSystemDefinition` and `parseGQLDocument` is replaced with `parseSchema`
+- `parseFullGQLDocument` replaced with `parseFullSchema`
 - removed `parseDSL` from `Data.Morpheus.Core`
 
 - following Types and modules are migrated to the new package `morpheus-graphql-app`: