+++ morpheus-graphql-core/changelog.md
@@ -1,12 +1,11 @@
 # Changelog
 
-## 0.17.0 - Unreleased
+## 0.17.0 - 25.02.2021
 
 ### New features
 
 - `Data.Morpheus.Core` provides default GrapHQL type definitions with `internalSchema`
 - exposed `Data.Morpheus.Internal.Ext`
-- exposed `Data.Morpheus.Types.GQLWrapper` with `EncodeWrapper` and `DecodeWrapper` type-classes.
 
 ### Breaking changes